Feature/standings after race

+++ F1WM/ApiModel/Standings/ConstructorPosition.cs
@@ -2,9 +2,9 @@ namespace F1WM.ApiModel
 {
 	public class ConstructorPosition
 	{
-		public int Id { get; set; }
+		public uint Id { get; set; }
 		public int Position { get; set; }
-		public int Points { get; set; }
+		public float Points { get; set; }
 		public ConstructorSummary Constructor { get; set; }
 	}
 }
