+++ F1WM/ApiModel/Standings/DriverPosition.cs
@@ -2,9 +2,9 @@ namespace F1WM.ApiModel
 {
 	public class DriverPosition
 	{
-		public int Id { get; set; }
+		public uint Id { get; set; }
 		public int Position { get; set; }
-		public double Points { get; set; }
+		public float Points { get; set; }
 		public DriverSummary Driver { get; set; }
 	}
 }
